==> pipeline/1-FeatureEngDrugs.ipynb <==
# Machine Learning Pipeline - Feature Engineering

# Paso 1: Reproducibility: Setting the seed

# to handle datasets
import pandas as pd
import numpy as np

# for plotting
import matplotlib.pyplot as plt

# for the yeo-johnson transformation
import scipy.stats as stats

# to divide train and test set
from sklearn.model_selection import train_test_split

# feature scaling
from sklearn.preprocessing import MinMaxScaler

# to save the trained scaler class
import joblib

# to visualise al the columns in the dataframe
pd.pandas.set_option('display.max_columns', None)

from pycaret.classification import *

# load dataset
data = pd.read_csv('./Drug.csv')

# rows and columns of the data
print(data.shape)

# visualise the dataset
data.head()

# Paso 2: Separamos el Dataset en entrenamiento y prueba

# Let's separate into train and test set
# Remember to set the seed (random_state for this sklearn function)

X_train, X_test, y_train, y_test = train_test_split(
    data.drop('Drug', axis=1), # predictive variables
    data['Drug'], # target
    test_size=0.1, # portion of dataset to allocate to test set
    random_state=0, # we are setting the seed here
shuffle=False)

X_train.shape, X_test.shape
print(X_train)
y_train= pd.DataFrame(y_train)
y_train.columns =['Drug']
y_test= pd.DataFrame(y_test)
y_test.columns =['Drug']
print(y_test)

# Paso 3: Feature Engineering (Ingeniería de Características)

En las siguientes celdas, procesaremos las variables del dataset de Precios de Casas para abordar:

1. Missing values (NaN - Valores Faltantes)
2. Temporal variables (Variables Temporales)
3. Non-Gaussian distributed variables (Variables con distribución no gaussiana)
4. Categorical variables: remove rare labels (Variables categóricas: eliminar etiquetas raras)
5. Categorical variables: convert strings to numbers (Variables categóricas: convertir cadenas a números)
5. Put the variables in a similar scale (Poner las variables en una escala similar)

## Numerical variable transformation - no aplica

### Logarithmic transformation


## Variables categóricas

### Aplicamos mapeos

# re-map strings to numbers

disc_mappings1 = {'LOW': 0, 'NORMAL': 1, 'HIGH': 2}

disc_vars1 = ['BP']

for var in disc_vars1:
    X_train[var] = X_train[var].map(disc_mappings1)
    X_test[var] = X_test[var].map(disc_mappings1)

disc_mappings2 = {'NORMAL': 0, 'HIGH': 1}

disc_vars2 = ['Cholesterol']

for var in disc_vars2:
    X_train[var] = X_train[var].map(disc_mappings2)
    X_test[var] = X_test[var].map(disc_mappings2)

disc_mappings3 = {'M':0, 'F':1}

disc_vars3 = ['Sex']

for var in disc_vars3:
    X_train[var] = X_train[var].map(disc_mappings3)
    X_test[var] = X_test[var].map(disc_mappings3)

    #Encoding de las variables categoricas a predecir.

disc_mappings4 = {'drugX':0, 'drugY':1, 'drugA':2, 'drugB':3,'drugC':4}
disc_vars4 = ['Drug']

for var in disc_vars4:
    y_train[var] = y_train[var].map(disc_mappings4)
    y_test[var] = y_test[var].map(disc_mappings4)


# check absence of na in the train set
[var for var in X_train.columns if X_train[var].isnull().sum() > 0]

#### Verificamos ¿Hay valores nulos (na) en el conjunto de Entrenamiento? - NO

[var for var in X_train.columns if X_train[var].isnull().sum() > 0]

#### Verificamos ¿Hay valores nulos (na) en el conjunto de Prueba? - NO

[var for var in X_test.columns if X_test[var].isnull().sum() > 0]

## Feature Scaling

Para su uso en modelos lineales, las características deben ser escaladas. Escalaremos las características a los valores mínimos y máximos:

if isinstance(y_train, pd.DataFrame):
    print("X_train es un DataFrame de pandas")
else:
    print("X_train no es un DataFrame de pandas")

print(len(X_train))
print(X_train.shape)

print(len(y_train))
print(y_train.shape)

print(X_test.shape)
print(y_test.shape)

# Creamos el Escalador (Scaler)
scaler = MinMaxScaler()

# Ajustamos el Scaler para el Conjunto de Entrenamiento
scaler.fit(X_train)

# Transformamos el conjunto de entrenamiento y el conjunto de prueba

# sklearn devuelve arrays de numpy, así que envolvemos el
# array con un dataframe de pandas

X_train = pd.DataFrame(
    scaler.transform(X_train),
    columns=X_train.columns
)

X_test = pd.DataFrame(
    scaler.transform(X_test),
    columns=X_train.columns
)

# X_train.head()
#print(X_train.shape)

# X_train
# y_train
print([X_train, y_train])

dataTrain= pd.concat([X_train, y_train], axis=1)
dataTrain
print(y_train.shape)
dataTest= pd.concat([X_test, y_test], axis=1)
print(y_test)
print("dataTest")
print(dataTest)

# Guardemos ahora los conjuntos de entrenamiento y prueba para el próximo Notebook!

X_train.to_csv('./OutputFeaturEngDrugs/xtrain.csv', index=False)
X_test.to_csv('./OutputFeaturEngDrugs/xtest.csv', index=False)

y_train.to_csv('./OutputFeaturEngDrugs/ytrain.csv', index=False)
y_test.to_csv('./OutputFeaturEngDrugs/ytest.csv', index=False)

# leemos ahora los conjuntos de entrenamiento y prueba para el próximo Notebook!

X_train= pd.read_csv('./OutputFeaturEngDrugs/xtrain.csv')
X_test= pd.read_csv('./OutputFeaturEngDrugs/xtest.csv')

y_train = pd.read_csv('./OutputFeaturEngDrugs/ytrain.csv')
y_test = pd.read_csv('./OutputFeaturEngDrugs/ytest.csv')

# Ahora guardamos el Scaler

joblib.dump(scaler, './OutputFeaturEngDrugs/minmax_scaler.joblib')

from pycaret.classification import setup

# Setting up the PyCaret environment with the training data and generating a profile report
clf1 = setup(data = dataTrain, target = 'Drug')

#Comparing all models 
best = compare_models()

evaluate_model(best)
#Picking the winner 
#best_model = automl(optimize = 'Accuracy')







#Fine-tuning the best model
#tuned_best_model = tune_model(best,n_iter=10, choose_better=True)

tuned_best_model



predictions = predict_model(tuned_best_model, data = dataTest)

predictions



from sklearn.metrics import accuracy_score
accuracy_score(predictions['Drug'], predictions['prediction_label'])

from sklearn.metrics import mean_squared_error, r2_score
import lightgbm as lgb

print("Loading data...")
# load or create your dataset
#regression_example_dir = Path(__file__).absolute().parents[1] / "regression"
#df_train = pd.read_csv(str(regression_example_dir / "regression.train"), header=None, sep="\t")
#df_test = pd.read_csv(str(regression_example_dir / "regression.test"), header=None, sep="\t")

#y_train = df_train[0]
#y_test = df_test[0]
#X_train = df_train.drop(0, axis=1)
#X_test = df_test.drop(0, axis=1)

print("Starting training...")
# train
gbm = lgb.LGBMRegressor(num_leaves=31, learning_rate=0.1, n_estimators=100)
gbm.fit(X_train, y_train, eval_set=[(X_test, y_test)], eval_metric="l1", callbacks=[lgb.early_stopping(5)])



print("Starting predicting...")
# predict
y_pred = gbm.predict(X_test, num_iteration=gbm.best_iteration_)




# eval
rmse_test = mean_squared_error(y_test, y_pred) ** 0.5
print(f"The RMSE of prediction is: {rmse_test}")

# feature importances
print(f"Feature importances: {list(gbm.feature_importances_)}")




# self-defined eval metric
# f(y_true: array, y_pred: array) -> name: str, eval_result: float, is_higher_better: bool
# Root Mean Squared Logarithmic Error (RMSLE)
def rmsle(y_true, y_pred):
    return "RMSLE", np.sqrt(np.mean(np.power(np.log1p(y_pred) - np.log1p(y_true), 2))), False


print("Starting training with custom eval function...")
# train
gbm.fit(X_train, y_train, eval_set=[(X_test, y_test)], eval_metric=rmsle, callbacks=[lgb.early_stopping(5)])




# another self-defined eval metric
# f(y_true: array, y_pred: array) -> name: str, eval_result: float, is_higher_better: bool
# Relative Absolute Error (RAE)
def rae(y_true, y_pred):
    return "RAE", np.sum(np.abs(y_pred - y_true)) / np.sum(np.abs(np.mean(y_true) - y_true)), False


print("Starting training with multiple custom eval functions...")
# train
gbm.fit(X_train, y_train, eval_set=[(X_test, y_test)], eval_metric=[rmsle, rae], callbacks=[lgb.early_stopping(5)])



print("Starting predicting...")
# predict
y_pred = gbm.predict(X_test, num_iteration=gbm.best_iteration_)
y_test2= y_test['Drug'].values
print(y_test2)
print(y_pred)



# eval
rmsle_test = rmsle(y_test2, y_pred)[1]
#rmsle_test = rmsle(1, 1.1)[1]
#rae_test = rae(y_test, y_pred)[1]
print(f"The RMSLE of prediction is: {rmsle_test}")
#print(f"The RAE of prediction is: {rae_test}")

# to build the model
from sklearn.linear_model import Lasso

# to evaluate the model
from sklearn.metrics import mean_squared_error, r2_score
import lightgbm as lgb

# set up the model
# remember to set the random_state / seed

#lin_model = Lasso(alpha=0.001, random_state=0)

# train the model

#lin_model.fit(X_train, y_train)




# make predictions for train set
pred = lin_model.predict(X_train)

# determine mse, rmse and r2
print('train mse: {}'.format(int(
    mean_squared_error(np.exp(y_train), np.exp(pred)))))
print('train rmse: {}'.format(int(
    mean_squared_error(np.exp(y_train), np.exp(pred), squared=False))))
print('train r2: {}'.format(
    r2_score(np.exp(y_train), np.exp(pred))))
print()

# make predictions for test set
pred = lin_model.predict(X_test)

# determine mse, rmse and r2
print('test mse: {}'.format(int(
    mean_squared_error(np.exp(y_test), np.exp(pred)))))
print('test rmse: {}'.format(int(
    mean_squared_error(np.exp(y_test), np.exp(pred), squared=False))))
print('test r2: {}'.format(
    r2_score(np.exp(y_test), np.exp(pred))))
print()

print('Average house price: ', int(np.exp(y_train).median()))

y_test.reset_index(drop=True)
