==> pipeline/1-PasosBasicosMejorModeloConPycaret.ipynb <==
# Probar Modelos con PyCaret (TEMPLATE DATAPATH - Prof. Inofuente)

# Paso 1: Cargamos Pycaret y un Dataset

import pycaret
pycaret.__version__

from pycaret.datasets import get_data
Datos_traffic = get_data(dataset = 'traffic') #Aquí cambias el Dataset que te haya tocado.

# Paso 2: Importar las Regresiones de Pycaret y Crear el Objeto

from pycaret.regression import *

# import RegressionExperiment and init the class
from pycaret.regression import RegressionExperiment
ExperimentoRegresion = RegressionExperiment() #Instanciamos la clase

# Paso 3: Configurar el Experimento u Objeto

# init setup on exp
ExperimentoRegresion.setup(Datos_traffic, target = 'traffic_volume', session_id = 123)

# Paso 4: Entrenamos y Comparamos los Modelos

#compare models using OOP
MejorModelo = ExperimentoRegresion.compare_models()

### *Resolver*
#### 1. ¿Cuál es el mejor modelo y por qué? Justificar. ¿Qué significa cada métrica?

print(MejorModelo)

ExperimentoRegresion.models()

compare_tree_models = ExperimentoRegresion.compare_models(include = ['dt', 'rf', 'et', 'gbr', 'xgboost', 'lightgbm', 'catboost'])

### *Resolver*
#### 2. ¿Por qué escogiste estos 7 modelos en la lista de include?. Si deseas puedes ayudarte del ordenamiento

# Paso 4 (Opcional):Ordenamiento de los mejores modelos respecto de una métrica

Mejores_Modelos_TOP3_MAE = ExperimentoRegresion.compare_models(sort = 'R2', n_select = 3)

# Paso 5: Tuneamos los hypermarámetros del Mejor Modelo

MejorModelo_Tuneado = tune_model(estimator = MejorModelo)

### *Resolver*
#### 3. ¿El mejor modelo tuneado ha tenido mejores valores de salida respecto a las métrica que el mejor modelo sin tunear?. Si tuvo o no, justifica tu respuesta.

# Paso 6: Finalizar Modelo

Modelo_Final = finalize_model(estimator = MejorModelo_Tuneado) #Escoger el MejorModelo o MejorModelo_Tuneado dependiendo de tu respuesta de arriba.

Modelo_Final

# Paso 7: Guardar Modelo

save_model(model = Modelo_Final, model_name = './Modelo_Final_1')

# Paso 8: Cargar Modelo (Hasta aquí entra la Tarea y la Evaluación)

Modelo_Cargado_desde_Disco = load_model('./Modelo_Final_1')

#### Predicciones (NO PROBAR Y NO ENTRA A LA TAREA NI A LA EVALUACIÓN - A PARTIR DE AQUÍ)

predictions = predict_model(Modelo_Cargado_desde_Disco , data=new_data)
predictions.head()
