==> pipeline/03-machine-learning-pipeline-feature-selection.ipynb <==
# Machine Learning Pipeline - Feature Selection

In this notebook, we pick up the transformed datasets that we saved in the previous notebook.

## Reproducibility: Setting the seed

With the aim to ensure reproducibility between runs of the same notebook, but also between the research and production environment, for each step that includes some element of randomness, it is extremely important that we **set the seed**.

# to handle datasets
import pandas as pd
import numpy as np

# for plotting
import matplotlib.pyplot as plt

# to build the models
from sklearn.linear_model import Lasso
from sklearn.feature_selection import SelectFromModel

# to visualise al the columns in the dataframe
pd.pandas.set_option('display.max_columns', None)

# load the train and test set with the engineered variables

# we built and saved these datasets in the previous lecture.
# If you haven't done so, go ahead and check the previous notebook
# to find out how to create these datasets

X_train = pd.read_csv('./Input/xtrain.csv')
X_test = pd.read_csv('./Input/xtest.csv')

X_train.head()

# load the target (remember that the target is log transformed)
y_train = pd.read_csv('./Input/ytrain.csv')
y_test = pd.read_csv('./Input/ytest.csv')

y_train.head()

### Feature Selection

Vamos a seleccionar un subconjunto de las características más predictivas. Hay un elemento de aleatoriedad en la regresión Lasso, así que recuerda establecer la semilla.

# Realizaremos el ajuste del modelo y la selección de características
# todo junto en unas pocas líneas de código

# primero, especificamos el modelo de Regresión Lasso y
# seleccionamos un alpha adecuado (equivalente a la penalización).
# Cuanto mayor sea el alpha, menos características serán seleccionadas.

# Luego usamos el objeto selectFromModel de sklearn, que
# seleccionará automáticamente las características cuyos coeficientes sean diferentes de cero

# recuerda establecer la semilla, el estado aleatorio en esta función
sel_ = SelectFromModel(Lasso(alpha=0.001, random_state=0))

# Entrenamos el Modelo Lasso y seleccionamos los features
sel_.fit(X_train, y_train)

sel_.get_support().sum()

# Visualicemos aquellas características que fueron seleccionadas.
# (características seleccionadas marcadas con True)

sel_.get_support()

# Imprimamos el número de características totales y seleccionadas

# Así es como podemos hacer una lista de las características seleccionadas
selected_feats = X_train.columns[(sel_.get_support())]

# let's print some stats
print('Número de Features en Total: {}'.format((X_train.shape[1])))
print('Número de Features Seleccionados: {}'.format(len(selected_feats)))
print('Características con coeficientes reducidos a cero: {}'.format(
    np.sum(sel_.estimator_.coef_ == 0)))

# print the selected features
selected_feats

pd.Series(selected_feats).to_csv('./Output/selected_features.csv', index=False)
